# Optimize `issymbolstart` and `issymbolcontinuation`